changes in notification route

+++ backend/controllers/friendRequest.js
@@ -14,20 +14,19 @@ router.get("/friendRequest/:id", validateUser,existingFriendOrNot, async(req,res
         if(!user){
          return res.status(400).send({message:"user not found"});
         }
-        console.log(id,userId);
         let friendObj= await Friend.create({sourceId:userId,targetId:id});
-
+        await friendObj.save();
  let notificationObj1={
-    friend:friendObj,
-    category:'sendRequest',
+    friend:friendObj._id,
+    category:"sendRequest",
     message:"send you a friend request."
 }
 await User.updateOne({_id:id,},{$push:{notifications:notificationObj1}});
 await user.save();
-res.status(201).send({message:"sent request"});
+return res.status(201).send({message:"success"});
 }
 catch(err){
-    res.status(500).send({message:err.message})
+  return res.status(500).send({message:err.message})
 }
 })
 
@@ -39,12 +38,10 @@ router.get("/acceptRequest/:index", validateUser, async(req,res)=>{
   let user = await User.findById(userId);
   
   if(user.notifications.length <= index){
-    return res.status(200).send({message:"no message in notification"});
+    return res.status(400).send({message:"empty notification"});
   }
   let friendId = user.notifications[index].friend.toString();
-  
   let friend= await Friend.findById(friendId);
-  console.log(friend);
   let senderId= friend.sourceId.toString();
   let targetId=friend.targetId.toString();
   let sender= await User.findById(senderId);
@@ -61,29 +58,32 @@ router.get("/acceptRequest/:index", validateUser, async(req,res)=>{
 sender.notifications.push(notificationObj);
    await user.save();
   await sender.save();
-  res.status(201).send({message:"accepted request"});
+  return res.status(201).send({message:"success"});
     }
     catch(err){
-        res.status(500).send({message:err.message})
+      return res.status(500).send({message:err.message})
     }
 })
 
 //reject request
 router.get("/rejectRequest/:index", validateUser ,async(req,res)=>{
     try{
-      let userId= req.user.id;
-       let {index}= req.params ;
-  let user= await User.findById(userId);
+      let userId = req.user.id;
+       let {index} = req.params ;
+  let user = await User.findById(userId);
+  if(user.notifications.length <= index){
+    return res.status(400).send({message:"empty notification"});
+  }
   let friendId = user.notifications[index].friend.toString();
-  
   let friend = await Friend.findByIdAndDelete(friendId);
-
-user.notifications.pop();
-user.save();
-  res.status(201).send({message:"friend Rrquest rejected"});
+  if(index >= 0){
+    user.notifications.splice(index,1);
+  }
+  await user.save();
+  return res.status(201).send({message:"success"});
     }
     catch(err){
-        res.status(500).send({message:err.message})
+      return res.status(500).send({message:err.message})
     }
 })
 
@@ -112,34 +112,24 @@ router.get("/getAllFriends",validateUser,async(req,res)=>{
       }
     }
     console.log(friendsList);
-
- 
-    res.status(201).send(friendsList);
+    return res.status(201).send(friendsList);
   }
   catch(err){
-   res.status(500).send({error:err.message});
+   return res.status(500).send({error:err.message});
   }
 });
 
-// delete notification
-router.delete("/notification/delete/:index", validateUser,async(req,res)=>{
-  try{
-
-    let userId= req.user.id;
-    let user= await User.findById(userId);
-    if(user.notifications.length<index){
-      res.status(400).send({message:"invalid request of deletion"}
-
-      )
-    }
-    user.notifications.splice(index,1);
-    user.save();
-    res.status(200).send({message:"notification as been deleted"})
+// check notification routes
+router.get("/user/notify", validateUser, async(req,res)=>{
+  try {
+    let userId = req.user.id;
+    console.log("user: ",userId);
+    let user = await User.findById(userId);
+    // let notification = user.notifications;
+    res.status(200).send({message:"success",notification:user});
+  } catch (error) {
+    res.status(500).send({message:err.message});
   }
-  catch(err){
-    res.status(500).send({message:err.message})
-  }
-
 })
